refactor(rss-reader): split FeedState and FeedAction into their own files
