+++ examples/rss-reader/shared/src/commonMain/kotlin/com/github/jetbrains/rssreader/app/FeedState.kt
@@ -1,0 +1,13 @@
+package com.github.jetbrains.rssreader.app
+
+import com.github.jetbrains.rssreader.domain.RssFeed
+
+data class FeedState(
+    val progress: Boolean,
+    val feeds: List<RssFeed>,
+    val selectedFeed: RssFeed? = null // null means selected all
+) : State
+
+fun FeedState.mainFeedPosts() =
+    (selectedFeed?.channel?.item ?: feeds.flatMap { it.channel?.item ?: emptyList() })
+        .sortedByDescending { it.pubDate }
